+++ src/main/java/com/teamddd/duckmap/DuckmapApplication.java
@@ -1,35 +1,16 @@
 package com.teamddd.duckmap;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @ConfigurationPropertiesScan
 @SpringBootApplication
 public class DuckmapApplication {
-	@Value("${props.allowed-origin-url}")
-	private String allowedFrontUrl;
 
 	public static void main(String[] args) {
 		SpringApplication.run(DuckmapApplication.class, args
 		);
-	}
-
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurer() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/**")
-					.allowedOrigins(
-						allowedFrontUrl,
-						"http://localhost:3000");
-			}
-		};
 	}
 
 }
